requirements.txt
setuptools==59.5.*
opencv-python
opencv-contrib-python
prettytable
tqdm>=4.64.0
yacs
torchfile
plyfile
munkres
chumpy
trimesh
shyaml
PyYAML==5.4.1
numpy>=1.22.0,<1.25.0
numpy-quaternion
pygame
keyboard
transforms3d
bvhtoolbox
vedo
imgaug==0.4.0
tensorboard>=2.10.0
torch>=1.12.0
smplx
pycocotools
plotly
open3d
addict
loguru
norfair
h5py==3.6.0
imageio-ffmpeg
pyrender
joblib
pandas==1.4.2
tensorflow==2.10.1
keras==2.10.0
scikit-learn==1.0.2
matplotlib==3.5.1
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.2
requests==2.27.1
pillow==9.0.1
concurrent-futures-extra==1.0.2
seaborn==0.11.2
pickle5